create helpers.js for helper functions
